requirements.txt
inspect_ai
ipykernel
openai
python-dotenv
tabulate